+++ internal/domain/errors.go
@@ -1,0 +1,5 @@
+package domain
+
+import "errors"
+
+var ErrNotFound = errors.New("not found")
