+++ internal/domain/room_member.go
@@ -1,0 +1,13 @@
+package domain
+
+import (
+	"time"
+
+	"github.com/google/uuid"
+)
+
+type RoomMember struct {
+	UserID          uuid.UUID
+	Name            string
+	LastConnectedAt time.Time
+}
