+++ internal/domain/user.go
@@ -1,0 +1,9 @@
+package domain
+
+import "github.com/google/uuid"
+
+type User struct {
+	ID     uuid.UUID
+	Name   string
+	APIKey string
+}
